requirements.txt
Django==2.2.6
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
